[bug] playerspawnevent
This originally used the playerJoinEvent and during this time we had to loop until the object, the player, loaded in before executing code to make sure that the player was properly targeted. The event executed prematurely before the player actually spawned so this left us will limited options to handle it. Fast forward, the API was introduced with a new event called playerSpawn and this event basically solved all of our problems. This was applied to Paradox but consequently the code itself was not modified to accomondate the new logic with this event. We want to make sure that this code is only execute during a players initial spawn. This means when they first join the world, and not execute with any other player spawns thereafter. This reduces the execution of the code a lot.
